WIP: keep cutover tests within LOC policy (checkpoint)
